Create signature for computer details used in C2
This is a conversion of this signature I was playing with: https://github.com/cuckoosandbox/community/pull/120 (I started with cuckoo 2.0 as I was doing a lot with signature conversion so I found it easier to start there and to mark calls).

Anyway this works with Dridex and Upatre currently from what I can see and I am looking for other samples I can use to build on this with. Currently it also only looks for username or computer name being used but again I am looking for more examples to build on with the intention of common bits of C2 information that may be sent about a computer that are good IOCs being targetted. 

The CryptoHashData call in Dridex doesn't show as cleanly as I hoped but it works fine. The intention though is to build upon this sig & cleaning up the reporting to hopefully get to a point where it can communicate some understanding of C2 beacon construction from bits of queried information (maybe once it is more complete and I can see what I am facing describing what bits of information has been used).