Update beacon sig with extra APIs
Following response on my original sig to detect sending data but was less effective than this signature from Killerinstinct would be following updates with these additional APIs. I have updated this signature with the logic of the other sig and checks so hopefully this will be suitable. 

Hopefully this signature could be expanded with other information about the computer which often appears in CnC & I will look at this after this update is looked at and ok'd.