Create signature for spawning/injecting many processes
This is a simple signature to detect malware with large process trees. One indicator for malware is lots of processes, this includes injection, creation or even running command windows etc as well as malware trying to be evasive (or rather annoying when viewing) by spawning large process trees from copies of itself. 

Anyway as mentioned simple and just take action based on the number of processes minus 1 for the parent process. Tested on various samples. I pondered whether or not to display the processes as appended information but decided against it as wasted processing time as user can click on behavior tab and see them anyway one this has been highlighted to them.

Also on a side note any feedback regarding the ransomware message submission now the suggested improvements have been made and tested as not had feedback since?